# Move geometric to Aten(CPU)