@XiaobingSuper 2020-01-09
Rebased to solve conflicts.